create separate Providers, made smth strange in Footer in handleClearLompleted
